+++ config/projection_test.go
@@ -388,4 +388,20 @@ func TestProjection(t *testing.T) {
 			}
 		})
 	})
+
+	t.Run("func Validate()", func(t *testing.T) {
+		t.Run("it panics if the source is present but the typename is not", func(t *testing.T) {
+			test.ExpectPanic(
+				t,
+				`malformed configuration representation: there is a problem with the *config.Projection value (and not necessarily the configuration it represents): Source is present, but TypeName is not`,
+				func() {
+					Validate(&Projection{
+						Source: optional.Some[dogma.ProjectionMessageHandler](
+							&ProjectionMessageHandlerStub{},
+						),
+					})
+				},
+			)
+		})
+	})
 }